[Refactor] Modify `tools/train.py` and some known bugs (#1582)
* modify tools/train.py

* fix some errors

* modify configs

* modify configs

* modify device of acc_pose

* fix unittests

* fix mpiidataset

* add save_best in configs

* add save_best in configs

+++ mmpose/codecs/msra_heatmap.py
@@ -133,7 +133,8 @@ class MSRAHeatmap(BaseKeypointCodec):
                     continue
 
                 # get gaussian center coordinates
-                mu = (keypoints[n, k] / self.scale_factor + 0.5)
+                mu = (keypoints[n, k] / self.scale_factor +
+                      0.5).astype(np.int64)
 
                 # check that the gaussian has in-bounds part
                 left, top = (mu - radius).astype(np.int64)
